📝 (everything): Updated comments to fit godoc standardxc
