+++ internal/node/node.go
@@ -3,11 +3,8 @@ package node
 import (
 	"container/list"
 	"context"
-	"strings"
 
-	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/sonr-io/core/internal/common"
-	"github.com/sonr-io/core/internal/device"
 	"github.com/sonr-io/core/internal/host"
 	"github.com/sonr-io/core/pkg/exchange"
 	"github.com/sonr-io/core/pkg/lobby"
@@ -78,25 +75,6 @@ func NewNode(ctx context.Context, host *host.SNRHost, loc *common.Location) *Nod
 	}
 	node.LobbyProtocol = lobby
 	return node
-}
-
-// Peer method returns the peer of the node
-func (n *Node) Peer() *common.Peer {
-	// Find PublicKey Buffer
-	pubBuf, err := crypto.MarshalPublicKey(n.host.PublicKey())
-	if err != nil {
-		logger.Error("Failed to marshal public key", zap.Error(err))
-		return nil
-	}
-
-	// Return Peer
-	return &common.Peer{
-		SName:     strings.ToLower(n.profile.SName),
-		Status:    common.Peer_ONLINE,
-		Device:    device.Info(),
-		Profile:   n.profile,
-		PublicKey: pubBuf,
-	}
 }
 
 // Edit method updates Node's profile
