--- pkg/exchange/state.go
@@ -1,1 +1,0 @@
-package exchange
